LayoutTests/ietestcenter/Javascript/15.2.3.3-4-186-expected.txt

15.2.3.3-4-186 (TestCases/chapter15/15.2/15.2.3/15.2.3.3/15.2.3.3-4-186.js) Object.getOwnPropertyDescriptor returns data desc (all false) for properties on built-ins (Function.length)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ES5Harness.preconditionPassed is true
FAIL ES5Harness.testPassed should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
